Made a continous-ish Julia fractal
Instead of running for a (potentially very large) number of iterations
and colouring according to how many iterations it took to get thrown
far away from the origin, this runs for a smallish number of iterations
and colours according to how far the current point got thrown away.

An early escape on that loop yields periodic colour bands at the edges
of the image, far away from the laser dots where the iteration is stable.