Set a hard max and min on state count
If the state count is set to 2, this thing flashes red and green, which looks
like it is probably an epilepsy hazard! Doesn't happen if it's 3 or higher.
Making the state count never go below 3 is safer.

If the state count is anywhere above about 30, the automaton is almost
guaranteed to get stuck. Around 250 some numerical bugs become apparent too, so
set a bound way below the point where the numerical bugs become visible. :)